refactor: rebase on master (although had to delete some files that were causing errors) & put DTOs, entity, and enums into domain package
