+++ backend/src/main/java/com/tcs/dhv/domain/dto/AddressRequestDto.java
@@ -1,4 +1,4 @@
-package com.tcs.dhv.dto;
+package com.tcs.dhv.domain.dto;
 
 import jakarta.validation.constraints.*;
