+++ backend/src/main/java/com/tcs/dhv/domain/dto/AddressResponseDto.java
@@ -1,4 +1,4 @@
-package com.tcs.dhv.dto;
+package com.tcs.dhv.domain.dto;
 
 public record AddressResponseDto(
         String line1,
