+++ backend/src/main/java/com/tcs/dhv/domain/dto/AddressUpdateDto.java
@@ -1,4 +1,4 @@
-package com.tcs.dhv.dto;
+package com.tcs.dhv.domain.dto;
 
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
