+++ backend/src/main/java/com/tcs/dhv/domain/dto/ParcelRequestDto.java
@@ -1,7 +1,7 @@
-package com.tcs.dhv.dto;
+package com.tcs.dhv.domain.dto;
 
-import com.tcs.dhv.enums.DeliveryType;
-import com.tcs.dhv.enums.PaymentType;
+import com.tcs.dhv.domain.enums.DeliveryType;
+import com.tcs.dhv.domain.enums.PaymentType;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
