+++ backend/src/main/java/com/tcs/dhv/domain/entity/Address.java
@@ -1,12 +1,11 @@
-package com.tcs.dhv.entity;
+package com.tcs.dhv.domain.entity;
 
-import jakarta.persistence.Column;
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -15,18 +14,18 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-
-import java.util.Date;
 import java.util.UUID;
+
 
 @Entity
 @Setter
 @Getter
 @ToString
 @Builder
-@AllArgsConstructor
 @NoArgsConstructor
-public class Recipient {
+@AllArgsConstructor
+@Table(name = "addresses")
+public class Address {
 
     @Id
     @Setter(AccessLevel.NONE)
@@ -34,20 +33,26 @@ public class Recipient {
     private UUID id;
 
     @NotNull
-    private String name;
+    private String line1;
+
+    private String line2;
+
+    private String building;
+
+    private String apartment;
 
     @NotNull
-    @Column(unique = true)
-    private String email;
-
-    private String phone;
-
-    private Date birthDate;
+    private  String city;
 
     @NotNull
-    @OneToOne
-    @JoinColumn(name = "address_id",nullable = false)
-    private Address address;
+    @Builder.Default
+    private String country = "Hungary";
+
+    @NotNull
+    private String postalCode;
+
+    private Double longitude;
+    private Double latitude;
 
 
 }
