+++ backend/src/main/java/com/tcs/dhv/domain/enums/LocationStatus.java
@@ -1,4 +1,4 @@
-package com.tcs.dhv.enums;
+package com.tcs.dhv.domain.enums;
 
 public enum LocationStatus {
     AVAILABLE,
