+++ backend/src/main/java/com/tcs/dhv/domain/enums/PaymentType.java
@@ -1,4 +1,4 @@
-package com.tcs.dhv.enums;
+package com.tcs.dhv.domain.enums;
 
 public enum PaymentType {
     SENDER_PAYS,
