+++ backend/src/main/java/com/tcs/dhv/repository/ParcelRepository.java
@@ -1,6 +1,6 @@
 package com.tcs.dhv.repository;
 
-import com.tcs.dhv.entity.Parcel;
+import com.tcs.dhv.domain.entity.Parcel;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
